added LT language to validation and more controller tests

+++ src/main/java/lt/swedbank/itacademy/carleasing/services/CorporateCustomerService.java
@@ -2,6 +2,7 @@ package lt.swedbank.itacademy.carleasing.services;
 
 import lt.swedbank.itacademy.carleasing.beans.documents.CorporateCustomer;
 import lt.swedbank.itacademy.carleasing.beans.responses.CorporateCustomerResponse;
+import lt.swedbank.itacademy.carleasing.exceptions.NotFoundException;
 import lt.swedbank.itacademy.carleasing.repositories.CorporateCustomerRepository;
 import org.bson.types.ObjectId;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -15,6 +16,7 @@ import java.util.stream.Collectors;
 @Service
 public class CorporateCustomerService {
 
+
     @Autowired
     private CorporateCustomerRepository repository;
 
@@ -22,7 +24,7 @@ public class CorporateCustomerService {
         return repository.findAll().stream().map(CorporateCustomerResponse::new).collect(Collectors.toList());
     }
 
-    public CorporateCustomer addNewCorporateCustomer(@Valid CorporateCustomer corporateCustomer) {
+    public CorporateCustomerResponse addNewCorporateCustomer(@Valid CorporateCustomer corporateCustomer) {
         CorporateCustomer newCorporateCustomer = new CorporateCustomer();
 
         newCorporateCustomer.setId(new ObjectId());
@@ -33,22 +35,35 @@ public class CorporateCustomerService {
         newCorporateCustomer.setCompanyName(corporateCustomer.getCompanyName());
         newCorporateCustomer.setPhoneNumber(corporateCustomer.getPhoneNumber());
 
-        return repository.save(newCorporateCustomer);
+        return new CorporateCustomerResponse(repository.save(newCorporateCustomer));
     }
 
     public CorporateCustomerResponse getCorporateCustomerById(String id) {
-        List<CorporateCustomerResponse> corporateCustomers = repository.findAll().stream().map(CorporateCustomerResponse::new).collect(Collectors.toList());
+        List<CorporateCustomerResponse> corporateCustomers = repository.findAll().stream()
+                .map(CorporateCustomerResponse::new)
+                .collect(Collectors.toList());
+
         for (CorporateCustomerResponse customer : corporateCustomers) {
             if (customer.getId().equals(id)) {
                 return customer;
             }
         }
-        return null;
-        //return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(null);
-        //throw new NullPointerException("404");
+        throw new NotFoundException("Sorry, but corporate customer with id: " + id + " do not present.");
     }
 
     public void deleteCorporateCustomer(String id) {
-        repository.delete(repository.findCorporateCustomersById(id));
+        List<CorporateCustomerResponse> corporateCustomers = repository.findAll().stream()
+                .map(CorporateCustomerResponse::new)
+                .collect(Collectors.toList());
+
+        for (CorporateCustomerResponse customer : corporateCustomers) {
+            if (customer.getId().equals(id)) {
+                repository.delete(repository.findCorporateCustomersById(id));
+            }
+            else{
+                throw new NotFoundException("Sorry, but corporate customer with id: " + id + " do not present.");
+            }
+        }
+
     }
 }
