+++ src/main/java/lt/swedbank/itacademy/carleasing/services/PrivateCustomerService.java
@@ -2,6 +2,7 @@ package lt.swedbank.itacademy.carleasing.services;
 
 import lt.swedbank.itacademy.carleasing.beans.documents.PrivateCustomer;
 import lt.swedbank.itacademy.carleasing.beans.responses.PrivateCustomerResponse;
+import lt.swedbank.itacademy.carleasing.exceptions.NotFoundException;
 import lt.swedbank.itacademy.carleasing.repositories.PrivateCustomerRepository;
 import org.bson.types.ObjectId;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -13,6 +14,7 @@ import java.util.stream.Collectors;
 @Service
 public class PrivateCustomerService {
 
+
     @Autowired
     private PrivateCustomerRepository repository;
 
@@ -21,16 +23,19 @@ public class PrivateCustomerService {
     }
 
     public PrivateCustomerResponse getPrivateCustomerById(String id) {
-        List<PrivateCustomerResponse> privateCustomers = repository.findAll().stream().map(PrivateCustomerResponse::new).collect(Collectors.toList());
+        List<PrivateCustomerResponse> privateCustomers = repository.findAll().stream()
+                .map(PrivateCustomerResponse::new)
+                .collect(Collectors.toList());
+
         for (PrivateCustomerResponse currentPrivateCustomer : privateCustomers) {
             if (currentPrivateCustomer.getId().equals(id)) {
                 return currentPrivateCustomer;
             }
         }
-        return null;
+        throw new NotFoundException("Sorry, but private customer with id: " + id + " do not present.");
     }
 
-    public PrivateCustomer addNewPrivateCustomer(PrivateCustomer privateCustomer) {
+    public PrivateCustomerResponse addNewPrivateCustomer(PrivateCustomer privateCustomer) {
         PrivateCustomer newPrivateCustomer = new PrivateCustomer();
 
         newPrivateCustomer.setId(new ObjectId());
@@ -42,11 +47,23 @@ public class PrivateCustomerService {
         newPrivateCustomer.setPersonalCode(privateCustomer.getPersonalCode());
         newPrivateCustomer.setPhoneNumber(privateCustomer.getPhoneNumber());
 
-        return repository.save(newPrivateCustomer);
+        return new PrivateCustomerResponse(repository.save(newPrivateCustomer));
     }
 
 
-    public void deletePrivateCustomer(String id) {
-        repository.delete(repository.findPrivateCustomerById(id));
+    public void removePrivateCustomer(String id){
+        List<PrivateCustomerResponse> corporateCustomers = repository.findAll().stream()
+                .map(PrivateCustomerResponse::new)
+                .collect(Collectors.toList());
+
+        for (PrivateCustomerResponse customer : corporateCustomers) {
+            if (customer.getId().equals(id)) {
+                repository.delete(repository.findPrivateCustomerById(id));
+            }
+            else {
+                throw new NotFoundException("Sorry, but private customer with id: " + id + " do not present.");
+            }
+        }
+
     }
 }
